Task 49 Code refactoring for services
